day 2 half of finished

+++ src/Order.h
@@ -15,6 +15,6 @@ struct Order {
     uint64_t id;
     Side side;
     OrderType type;
-    double price; // For LIMIT orders, this is the limit price. For MARKET orders, this can be ignored.
+    int price; // For LIMIT orders, this is the limit price. For MARKET orders, this can be ignored.
     uint32_t quantity; // The quantity of the order.
 };
